requirements.txt
absl-py==0.7.1
altair==3.2.0
appnope==0.1.0
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.7.1
blis==0.4.1
catalogue==1.0.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.3
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
dask==2.8.0
decorator==4.4.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.4.1
fsspec==0.6.0
future==0.18.2
gast==0.2.2
get==2019.4.13
grpcio==1.20.1
gunicorn==20.0.4
h5py==2.9.0
idna==2.8
ijson==2.6.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jieba3k==0.35.1
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.1.1
kiwisolver==1.1.0
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
newspaper3k==0.2.8
notebook==6.0.1
npm==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plac==1.1.3
post==2019.4.13
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.7.1
psycopg2
ptyprocess==0.6.0
public==2019.4.13
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.0
python-dotenv
pytz==2019.1
PyYAML==5.1
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
regex==2019.6.8
request==2019.4.13
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
retrying==1.3.3
sacrebleu==1.3.6
scikit-learn==0.22
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
spacy==2.2.3
SQLAlchemy==1.3.17
srsly==1.0.1
statsmodels==0.10.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
tinysegmenter==0.3
tldextract==2.2.1
toolz==0.10.0
tqdm==4.32.2
traitlets==4.3.3
tweepy==3.8.0
typing==3.7.4
Unidecode==1.1.1
urllib3==1.25.3
vega==2.6.0
vega-datasets==0.7.0
virtualenv==16.6.0
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wget==3.2
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==0.6.0
